Split up vcs tests
