+++ tests/vcs/test_clone.py
@@ -1,106 +1,8 @@
 # -*- coding: utf-8 -*-
-
-"""
-test_vcs
-------------
-
-Tests for `cookiecutter.vcs` module.
-"""
-
 import os
 import pytest
 
 from cookiecutter import exceptions, vcs
-
-
-@pytest.mark.parametrize('repo_url, exp_repo_type, exp_repo_url', [
-    (
-        "git+https://github.com/pytest-dev/cookiecutter-pytest-plugin.git",
-        "git",
-        "https://github.com/pytest-dev/cookiecutter-pytest-plugin.git"
-    ), (
-        "hg+https://bitbucket.org/foo/bar.hg",
-        "hg",
-        "https://bitbucket.org/foo/bar.hg"
-    ), (
-        "https://github.com/pytest-dev/cookiecutter-pytest-plugin.git",
-        "git",
-        "https://github.com/pytest-dev/cookiecutter-pytest-plugin.git"
-    ), (
-        "https://bitbucket.org/foo/bar.hg",
-        "hg",
-        "https://bitbucket.org/foo/bar.hg"
-    )
-])
-def test_identify_known_repo(repo_url, exp_repo_type, exp_repo_url):
-    assert vcs.identify_repo(repo_url) == (exp_repo_type, exp_repo_url)
-
-
-@pytest.fixture(params=[
-    "foo+git",  # uses explicit identifier with 'git' in the wrong place
-    "foo+hg",  # uses explicit identifier with 'hg' in the wrong place
-    "foo+bar",  # uses explicit identifier with neither 'git' nor 'hg'
-    "foobar"  # no identifier but neither 'git' nor 'bitbucket' in url
-])
-def unknown_repo_type_url(request):
-    return request.param
-
-
-def test_identify_raise_on_unknown_repo(unknown_repo_type_url):
-    with pytest.raises(exceptions.UnknownRepoType):
-        vcs.identify_repo(unknown_repo_type_url)
-
-
-def test_prompt_should_ask_and_rm_repo_dir(mocker, tmpdir):
-    """In `prompt_and_delete_repo()`, if the user agrees to delete/reclone the
-    repo, the repo should be deleted.
-    """
-    mock_read_user = mocker.patch(
-        'cookiecutter.vcs.read_user_yes_no',
-        return_value=True,
-        autospec=True
-    )
-    repo_dir = tmpdir.mkdir('repo')
-
-    vcs.prompt_and_delete_repo(str(repo_dir))
-
-    assert mock_read_user.called
-    assert not repo_dir.exists()
-
-
-def test_prompt_should_ask_and_keep_repo_dir(mocker, tmpdir):
-    """In `prompt_and_delete_repo()`, if the user wants to keep their old
-    cloned template repo, it should not be deleted.
-    """
-    mock_read_user = mocker.patch(
-        'cookiecutter.vcs.read_user_yes_no',
-        return_value=False,
-        autospec=True
-    )
-    repo_dir = tmpdir.mkdir('repo')
-
-    with pytest.raises(SystemExit):
-        vcs.prompt_and_delete_repo(str(repo_dir))
-
-    assert mock_read_user.called
-    assert repo_dir.exists()
-
-
-def test_prompt_should_not_ask_if_no_input_and_rm_repo_dir(mocker, tmpdir):
-    """In `prompt_and_delete_repo()`, if `no_input` is True, the call to
-    `vcs.read_user_yes_no()` should be suppressed.
-    """
-    mock_read_user = mocker.patch(
-        'cookiecutter.vcs.read_user_yes_no',
-        return_value=True,
-        autospec=True
-    )
-    repo_dir = tmpdir.mkdir('repo')
-
-    vcs.prompt_and_delete_repo(str(repo_dir), no_input=True)
-
-    assert not mock_read_user.called
-    assert not repo_dir.exists()
 
 
 @pytest.fixture
